livekit/02-onboarding-agent/pyproject.toml
[project]
name = "assemblyai-livekit-onboarding"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "livekit-agents[silero,turn-detector]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv",
    "aiohttp",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
